learn vue 6

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,11 @@ use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -66,6 +71,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+//        if (\Gate::denies('update-question', $question)) {
+//            abort(403, 'Access Denied');
+//        }
+        $this->authorize('update', $question);
         return view('questions.edit', compact('question'));
     }
 
@@ -78,6 +87,10 @@ class QuestionsController extends Controller
      */
     public function update(Request $request, Question $question)
     {
+//        if (\Gate::denies('update-question', $question)) {
+//            abort(403, 'Access Denied');
+//        }
+        $this->authorize('update', $question);
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', "Your question has been updated.");
@@ -91,6 +104,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+//        if (\Gate::denies('delete-question', $question)) {
+//            abort(403, 'Access Denied');
+//        }
+        $this->authorize('delete', $question);
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted.");
